errors: Lop off the string part of BorshIoError
#### Problem

The `BorshIoError` variant of `InstructionError` was changed from a
string to a unit variant in the Rust SDK in
[this PR](https://github.com/anza-xyz/solana-sdk/pull/12), but the
change hasn't been reflected in Kit yet.

#### Summary of changes

Change the `BorshIoError` variant to lop off any string that comes back
for compatibility with old and new errors emitted by the runtime.

Add the new variant to the RPC error type list.